refactor: rename config reconciler package
